Remove startup.sh
This is specific to our fork of MapIt, and was never mentioned in
the README and other docs. It assumes the developer has 'virtualenv'
installed locally, but there are no instructions for this anywhere.
Removing it is consistent with our approach for other repos, and we
expect developers will use GOV.UK Docker in the first instance.