Remove the need to download then upload then download files
The instructions currently require us to download two files (around
500MB and 200MB respectively), upload them to S3 then download them
again when running the import.

The files on S3 are not used for any other purpose.

This is counterproductive and time consuming when on a slow internet
connection.

Therefore updating so we can use a local file, so there is only one
download required.